+++ benchmark/benches/bench_world_add_remove_8_soa.jl
@@ -1,15 +1,15 @@
 
 function setup_world_add_remove_8_soa(n_entities::Int)
     world = World(
-        Position => StructArrayStorage,
-        CompN{1} => StructArrayStorage,
-        CompN{2} => StructArrayStorage,
-        CompN{3} => StructArrayStorage,
-        CompN{4} => StructArrayStorage,
-        CompN{5} => StructArrayStorage,
-        CompN{6} => StructArrayStorage,
-        CompN{7} => StructArrayStorage,
-        CompN{8} => StructArrayStorage,
+        Position => Storage{StructArray},
+        CompN{1} => Storage{StructArray},
+        CompN{2} => Storage{StructArray},
+        CompN{3} => Storage{StructArray},
+        CompN{4} => Storage{StructArray},
+        CompN{5} => Storage{StructArray},
+        CompN{6} => Storage{StructArray},
+        CompN{7} => Storage{StructArray},
+        CompN{8} => Storage{StructArray},
     )
 
     entities = Vector{Entity}()