+++ src/Ark.jl
@@ -57,6 +57,8 @@ export OnAddRelations, OnRemoveRelations
 export Observer, observe!, emit_event!
 export unpack, @unpack
 
-export StructArrayStorage, VectorStorage, Relationship
+export Relationship
+
+export Storage, StructArray
 
 end