+++ src/fieldsview.jl
@@ -6,7 +6,8 @@ end
 """
     unpack(a::FieldViewable)
 
-Unpacks the components (i.e. field vectors) of a [VectorStorage](@ref) column returned from a [Query](@ref).
+Unpacks the components (i.e. field vectors) of a column returned from a [Query](@ref)
+when the storage is different from `Storage{StructArray}`.
 See also [@unpack](@ref).
 
 !!! note