+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Breaking changes
 
+- Storage modes wrap the storage types instead of using unrelated types (#437)
 - Batch entity creation uses a callback/`do`-block instead of an iterator (#400)
 
 ### Features
@@ -15,6 +16,7 @@
 - Adds batch entity removal (#396, #402)
 - Adds batch-setting entity relation targets (#406)
 - Adds batched versions of adding, removing and exchanging components (#408)
+- Adds support for user-defined storage types (#437)
 - Makes Ark compatible with Mooncake (#405)
 
 ### Performance